% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_FElasso.R
\name{big.FE.lasso}
\alias{big.FE.lasso}
\title{Perform LASSO for feature selection and create linear
regression using selected features using a big.matrix}
\usage{
big.FE.lasso(X, yvar = "incidence", labelvar = c("cancer", "age",
  "Country_Transco", "year", "area.x", "area.y"), crossvalidation = T,
  returnplot = T, nfolds = 10, ncores = 1)
}
\arguments{
\item{X}{A \code{big.matrix} object}

\item{yvar}{Column name for the explained variable}

\item{labelvar}{Column name for additional variables}

\item{crossvalidation}{Should we perform cross-validation}

\item{nfolds}{Number of folds for cross validation.
Ignored if \code{crossvalidation = F}. If \code{crossvalidation}
equals the number of observations, a leave-one-out cross validation
is performed}

\item{ncores}{Number of cores used for computations. If \code{ncores}>1,
parallel processing is used}
}
\value{
Linear model fitted with selected features
}
\description{
Perform LASSO for feature selection and create linear
regression using selected features using a big.matrix
}
