% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_factor.R
\name{check.factor}
\alias{check.factor}
\title{Determine if a variable should be of factor class}
\usage{
check.factor(data, check.levels = F, threshold = 10)
}
\arguments{
\item{data}{A dataframe}

\item{check.levels}{Boolean indicating whether we
want to transform dataset (\code{FALSE}) or only
return number of levels factor variables would have
if converted (\code{TRUE})}
}
\value{
A dataframe where some columns have been converted
to factor class or a named vector where potential factor
variables are associated with number of levels they would
have if converted
}
\description{
Determine if a variable should be of factor class
}
\examples{
\dontrun{
check.factor(df_FAO,check.levels = T)
df <- check.factor(df_FAO)}
}
