% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_falsezero.R
\name{convert_fakezero}
\alias{convert_fakezero}
\title{Detect potential miscoding NAs}
\usage{
convert_fakezero(df, convert = F)
}
\description{
\code{convert_fakezero} is designed to inspect
a dataframe and detect potential miscoded NAs as 0.
The suspected miscoded columns
are returned if \code{convert = F} while suspects are
converted if \code{convert = T}
}
\examples{
\dontrun{
# ASSUMING YOU HAVE A df_ILO DATAFRAME IN ENVIRONMENT
convert_fakezero(df)
convert_fakezero(df,convert=T)}
}
