% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_NAs.R
\name{count_NAs}
\alias{count_NAs}
\title{Count share of missing values for all variables by group}
\usage{
count_NAs(df, groupingvar = "Country_Transco")
}
\arguments{
\item{df}{Dataframe}

\item{groupingvar}{Character vector indicating variable
names used for grouping (e.g. country level variable)}
}
\value{
A dataframe where, for each group, the proportion
of missing values is returned for each column. With a
\code{n x p} dataframe, \code{count_NAs} returns an
\code{np x 3} dataframe
}
\description{
Count share of missing values for all variables by group
}
\examples{
\dontrun{
# ASSUMING YOU HAVE A df_ILO DATAFRAME IN ENVIRONMENT
count_NAs(df)
count_NAs(df, "year")}
}
\seealso{
\code{check.emptycolumn}
}
