% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_character.R
\name{encode.character}
\alias{encode.character}
\title{Encode character vector with numeric values}
\usage{
encode.character(df, char.var = "Country_Transco", codelist = T)
}
\arguments{
\item{df}{Dataframe}

\item{char.var}{Character variable name that should be converted}

\item{codelist}{If \code{FALSE}, a numeric vector is returned
with numeric codes. If \code{TRUE}, a dataframe is returned indicating
the correspondance between the numeric values and the initial character
vector}
}
\value{
A dataframe of two columns or a numeric vector
}
\description{
\code{encode.character} takes a character column of a dataframe
and returns it converted with numeric values. It is also possible
to return the correspondance between the numeric values and the
initial characters
}
