% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_data_complete.R
\name{interpolate_group_complete}
\alias{interpolate_group_complete}
\title{Given a grouped dataframe and a line, interpolate numeric variables and impute missing data from non numeric variables with an extended locf preserving the dimension}
\usage{
interpolate_group_complete(missingdata, df, i = 1, year.var = "year")
}
\arguments{
\item{missingdata}{A dataframe listing groups and variables
that must be changed}

\item{i}{Line index of missingdata that will be looked}
}
\value{
A list with several components: \code{lines} and \code{cols}
return the lines and columns concerned by imputation, \code{after}
returns the relevant lines and columns imputed while \code{change}
stores as a dummy whether data have been imputed with interpolation or extended locf.
}
\description{
Given a grouped dataframe and a line, interpolate numeric variables and impute missing data from non numeric variables with an extended locf preserving the dimension}
\details{
Intermediate function, do not use
}
\examples{
# See \\code{interpolate_data_complete} source code
}
