% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performPreProcess.R
\name{performPreProcess}
\alias{performPreProcess}
\title{Preprocess data to harmonize scales}
\usage{
performPreProcess(X, groupingvar = c("sex", "age", "Country_Transco"),
  labelvar = c("year", "Zonier"))
}
\arguments{
\item{X}{A \code{big.matrix} object}

\item{groupingvar}{Variables used for grouping observations}

\item{labelvar}{Additional variables that should not be
pre-processed}
}
\value{
A \code{preProcess} object
}
\description{
Preprocess data to harmonize scales
}
